requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.28   
google-genai>=0.3.0
edge-tts>=6.1.10
httpx>=0.27.0
ffmpeg-python>=0.2.0
jinja2>=3.1.3        
python-multipart>=0.0.9